Major refactor of Token to remove generics.
Nearly every file was touched, in a lot of cases just replacing generic bounds.

TokenListCloner and the lowercase preprocessor removed as both were very
dependent on the old-style Tokens

New tokens maintain a state of what type of token they are. A list class was
added specifically for them which ensures that only tokens of a given type can
be contained within the list. This new list implements the functionality of the
old TokenListCloner.

Removed Smith-Waterman tests, which will be rewritten in the future - they were
not particularly comprehensive or good anyways.

Signed-off-by: Matthew Heon <matthew.heon@gmail.com>