Major changes to ChecksimsConfig
The configuration class has been refactored to move to a builder pattern, to
remove the need for one gigantic constructor while retaining immutability.

ChecksimsConfig now contains the actual submissions Checksims is intended to
operate on, not the directories in which they are contained. This opens up the
possibility of alternative modes of specifying submissions, including
submissions not from the same directories.

PreprocessSubmissions and RemoveCommonCode now operate on Collections instead of
lists in their APIs - they should not provide a guarantee of ordering, so
Collection is more appropriate.

Signed-off-by: Matthew Heon <matthew.heon@gmail.com>