Implement a new threading model to replace parallel streams
Parallel streams use the dispatching thread as a compute thread as well. This
means that, when the dispatching thread is blocked computing similarities (as it
often will be!), other threads that complete their jobs will idle. Swap to a
thread pool we manage ourselves to fix this.

At present, only used for actual algorithm computation, not for common code
removal (which still uses parallel streams). Preprocessors are not parallel at
all at present.

Signed-off-by: Matthew Heon <matthew.heon@gmail.com>