Parallelize Common Code Removal with new framework
The new parallel execution framework has been made generic to enable this to
occur, and to make adding further parallel operations easier.

There's also a grabbag of small changes sprinkled in here. Unit tests for Line
Compare, more TokenList unit tests, a bugfix for Line Compare when called on two
of the same submission, and other smaller changes.

There are the beginnings of unit tests for common code removal, which fail. It
appears that common code removal is Thoroughly Bugged at the moment. Future
commits will attempt to rectify this, and a possible race condition in token
creation.

Signed-off-by: Matthew Heon <matthew.heon@gmail.com>