Add new implementation of Smith-Waterman algorithm and unit tests
There are definitely still bugs with this implementation, but it actually runs
without causing critical errors, which is a definite plus.

Produces results which can vary from the previous implementation by around 5% in
real-world use. This could be due to the implementation bugs in the earlier
version, but I'm unsure as of right now.

Should be slightly faster than previous implementation, but needs yet more
tuning and optimizations - there's some low-hanging fruit that should be able
to significantly accelerate detection.

Signed-off-by: Matthew Heon <matthew.heon@gmail.com>